+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## master / unreleased
 
-* [FEATURE] promtail positions file corruptions can be ignored with the `positions.ignore-corruptions` flag. In the case the positions yaml is corrupted an empty positions config will be used and should later overwrite the malformed yaml.
+* [FEATURE] promtail positions file corruptions can be ignored with the `positions.ignore-invalid-yaml` flag. In the case the positions yaml is corrupted an empty positions config will be used and should later overwrite the malformed yaml.
 
 # 1.2.0 (2019-12-09)